fix: Extracted logic of calculating defaultStepValue into custom hook. Simplified implementation of getDefaultStepValue function
